COMPONENTS & PROPS
Add a method within a child coponent which emits or triggers an event
within child component that the parent components listens for, to
trigger a method on the parent component.

In this case a Add Student button was placed within a child component.
Once clicked, an event the parent component is listening for is
triggered, which increments the value of student count by 1 on the
parent component.